+++ burger-queen/src/components/admin/adminUser/infoUser.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { deleteUser } from "../../../petitions/userPetition";
 import { BurgerContext } from "../../../context/indexContext";
+import Swal from 'sweetalert2'
 
 
 const UserItem = (props) => {
@@ -16,11 +17,28 @@ const UserItem = (props) => {
     } = React.useContext(BurgerContext);
 
     const deleteUserBtn = () => {
-        deleteUser(props.id, props)
-        console.log(users);
-        // filtra el estado para volverlo a renderizar
+        Swal.fire({
+            title: 'Estás seguro?',
+            text: "No podrás recuperar el usuario!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText:  'Sí, eliminar!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+                deleteUser(props.id, props)
+                 // filtra el estado para volverlo a renderizar
         const arrayFilterUser = users.filter(user => user.id !== props.id )
-       setUsers(arrayFilterUser)
+        setUsers(arrayFilterUser)
+              Swal.fire(
+                'Eliminado!',
+                            'El Usuario ha sido eliminada correctamente',
+                            'Eliminada correctamente',
+              )
+            }
+          })
+
     }
 
     const saveUserEdit = (e) => {
